bug: use sentry 0.19 to match our backend server  (#276)
Closes #275

+++ autopush-common/src/db/mod.rs
@@ -412,7 +412,7 @@ impl DynamoStorage {
                 self.metrics.clone(),
                 table_name,
                 uaid,
-                11 as u32,
+                11,
             ))
         } else {
             Box::new(future::ok(Default::default()))
@@ -446,7 +446,7 @@ impl DynamoStorage {
                         table_name.as_ref(),
                         &uaid,
                         timestamp,
-                        10 as u32,
+                        10,
                     ))
                 } else {
                     Box::new(future::ok(Default::default()))
